register read_url tool
